allow for spaces in class names
If spaces are forbidden in class names, there is no way to declare multiple
classes (such as 'span3' and 'pull-right') for elements. Using
"p(span3)(pull-right)." will lead to <p class="span3" class="pull-right">,
which is definitely not right. On the other hand, allowing for spaces in class
names does not seem to impose a security problem.