Deprecation fixes for when we can drop 2.9.
One XmlElem.apply has been deprecated in favour of an alternative introduced in 2.10.  This branch contains removal of the deprecations but cannot be deployed until we drop 2.9.

The change is to minimise tags if empty. E.g., <p></p> to <p/>.  For the most part this is false in the code, or a test on the element.isEmpty.